Add files via upload
The analysis in this report focuses on identifying the key player attributes that contribute to both offensive and defensive success in football. We explored various player metrics, such as Finishing, Dribbling, Marking, and Tackling, to determine their impact on player performance and roles on the field. The analysis also differentiates between offensive attributes, such as Positioning and Finishing, which are crucial for goal-scoring, and defensive attributes like Marking and Interceptions, which are essential for stopping attacks. Additionally, we assessed the Overall and Potential ratings of players to understand their current abilities and future growth potential. The role of physical attributes, such as Stamina and Strength, was examined to highlight their importance in sustaining performance throughout matches. Overall, the analysis provides data-driven insights that help in optimizing team composition, assigning roles based on players’ strengths, and informing training and development strategies for both immediate and long-term team success.